Greatly reduced the number of PortDisplay classes
PortDisplayInput and PortDisplayOutput now replace 11 classes and removes the complex class hierarchy entirely.
Now ConduitType is an argument and no longer set by selecting the correct class
The classes were used to allow the computer to catch if a port was connected to the wrong overlay, a problem, which no longer exist.
The introduction of caches have moved overlay selection to NightLib, meaning the building modder no longer sets the overlay.